requirements.txt
python-telegram-bot==21.3
google.generativeai
pydub
google-cloud-storage
google-cloud-secret-manager
numpy
scipy
openai-whisper